fix to read/write data on big-endian machine
